+++ src/components/dialog/dialog.styles.ts
@@ -13,7 +13,7 @@ export default css`
     display: contents;
   }
 
-  .lynk-dialog {
+  .dialog {
     display: flex;
     align-items: center;
     justify-content: center;
@@ -25,7 +25,7 @@ export default css`
     z-index: var(--lynk-z-index-dialog);
   }
 
-  .lynk-dialog__panel {
+  .dialog__panel {
     display: flex;
     flex-direction: column;
     z-index: 2;
@@ -38,28 +38,28 @@ export default css`
     transition: var(--lynk-transition-fast) width, var(--lynk-transition-fast) height;
   }
 
-  .lynk-dialog__panel:focus {
+  .dialog__panel:focus {
     outline: none;
   }
 
   /* Ensure there's enough vertical padding for phones that don't update vh when chrome appears (e.g. iPhone) */
   @media screen and (max-width: 420px) {
-    .lynk-dialog__panel {
+    .dialog__panel {
       max-height: 80vh;
     }
   }
 
-  .lynk-dialog--open .lynk-dialog__panel {
+  .dialog--open .dialog__panel {
     display: flex;
     opacity: 1;
   }
 
-  .lynk-dialog__header {
+  .dialog__header {
     flex: 0 0 auto;
     display: flex;
   }
 
-  .lynk-dialog__title {
+  .dialog__title {
     flex: 1 1 auto;
     font: inherit;
     font-size: var(--lynk-font-size-large);
@@ -68,7 +68,7 @@ export default css`
     margin: 0;
   }
 
-  .lynk-dialog__header-actions {
+  .dialog__header-actions {
     flex-shrink: 0;
     display: flex;
     flex-wrap: wrap;
@@ -77,15 +77,15 @@ export default css`
     padding: 0 var(--header-spacing);
   }
 
-  .lynk-dialog__header-actions lynk-icon-button,
-  .lynk-dialog__header-actions ::slotted(lynk-icon-button) {
+  .dialog__header-actions lynk-icon-button,
+  .dialog__header-actions ::slotted(lynk-icon-button) {
     flex: 0 0 auto;
     display: flex;
     align-items: center;
     font-size: var(--lynk-font-size-medium);
   }
 
-  .lynk-dialog__body {
+  .dialog__body {
     flex: 1 1 auto;
     display: block;
     padding: var(--body-spacing);
@@ -93,21 +93,21 @@ export default css`
     -webkit-overflow-scrolling: touch;
   }
 
-  .lynk-dialog__footer {
+  .dialog__footer {
     flex: 0 0 auto;
     text-align: right;
     padding: var(--footer-spacing);
   }
 
-  .lynk-dialog__footer ::slotted(lynk-button:not(:first-of-type)) {
+  .dialog__footer ::slotted(lynk-button:not(:first-of-type)) {
     margin-inline-start: var(--lynk-spacing-x-small);
   }
 
-  .lynk-dialog:not(.lynk-dialog--has-footer) .lynk-dialog__footer {
+  .dialog:not(.dialog--has-footer) .dialog__footer {
     display: none;
   }
 
-  .lynk-dialog__overlay {
+  .dialog__overlay {
     position: fixed;
     top: 0;
     right: 0;
@@ -117,7 +117,7 @@ export default css`
   }
 
   @media (forced-colors: active) {
-    .lynk-dialog__panel {
+    .dialog__panel {
       border: solid 1px var(--lynk-color-neutral-0);
     }
   }
@@ -126,12 +126,12 @@ export default css`
   * Size modifiers
   */
 
-  .lynk-dialog--fitted .lynk-dialog__panel {
+  .dialog--fitted .dialog__panel {
     width: 80vw;
     min-height: 80vh;
   }
 
-  .lynk-dialog--full .lynk-dialog__panel {
+  .dialog--full .dialog__panel {
     width: calc(100% - var(--lynk-spacing-2x-large));
     height: calc(100% - var(--lynk-spacing-2x-large));
   }